caveats: inherit helpers from Hbc::DSL::Base
